code/requirements.txt
pandas
folium
sqlalchemy
PublicDataReader
PyYAML
openpyxl
streamlit 
